cooked prototype
very cooked prototype. you click on a calendar item and it moves you to a classes screen. It should only be clickable for the days marked on the calendar but idk what got cooked up. At least it goes to a new screen when you click on a day. I've explained a little bit of how we could get it working properly in the comments.