"forum" + name change + sending data to classbar + kind of fixed the text input stuff
theres a "forum" for each lesson now but we need a better lesson id system. rn it identifies the lesson from which to pull the forum from by the class name and date, but each class really neds a unique id we can get from the name. rn the class information is passed to the default screen, which passes it to the bar, which can pass it to any other screen (including documents). added padding to the text inputs for the asssistant and forum TODO: Documents screen, make text input boxes move with the keyboard (so the user can see what theyre typing)